backend/requirements.txt
psycopg2-binary
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
sentence-transformers
numpy
python-dotenv
pgvector
openai